+++ merlin/models/tf/__init__.py
@@ -29,32 +29,6 @@ from tensorflow.python.keras.optimizer_v2.optimizer_v2 import OptimizerV2
 from tensorflow.python.training.tracking.data_structures import ListWrapper, _DictWrapper
 
 from merlin.models.loader.tf_utils import configure_tensorflow
-from merlin.models.tf.blocks.core.aggregation import (
-    ConcatFeatures,
-    ElementwiseSum,
-    ElementwiseSumItemMulti,
-    StackFeatures,
-)
-from merlin.models.tf.blocks.core.base import (
-    Block,
-    EmbeddingWithMetadata,
-    ModelContext,
-    NoOp,
-    right_shift_layer,
-)
-from merlin.models.tf.blocks.core.combinators import ParallelBlock, ResidualBlock, SequentialBlock
-from merlin.models.tf.blocks.core.index import IndexBlock, TopKIndexBlock
-from merlin.models.tf.blocks.core.inputs import InputBlock
-from merlin.models.tf.blocks.core.masking import CausalLanguageModeling, MaskedLanguageModeling
-from merlin.models.tf.blocks.core.tabular import AsTabular, Filter, TabularBlock
-from merlin.models.tf.blocks.core.transformations import (
-    AsDenseFeatures,
-    AsSparseFeatures,
-    CategoricalOneHot,
-    ExpandDims,
-    LabelToOneHot,
-    StochasticSwapNoise,
-)
 from merlin.models.tf.blocks.cross import CrossBlock
 from merlin.models.tf.blocks.dlrm import DLRMBlock
 from merlin.models.tf.blocks.experts import CGCBlock, MMOEBlock, MMOEGate
@@ -74,6 +48,32 @@ from merlin.models.tf.blocks.sampling.cross_batch import (
 )
 from merlin.models.tf.blocks.sampling.in_batch import InBatchSampler
 from merlin.models.tf.blocks.sampling.queue import FIFOQueue
+from merlin.models.tf.core.aggregation import (
+    ConcatFeatures,
+    ElementwiseSum,
+    ElementwiseSumItemMulti,
+    StackFeatures,
+)
+from merlin.models.tf.core.base import (
+    Block,
+    EmbeddingWithMetadata,
+    ModelContext,
+    NoOp,
+    right_shift_layer,
+)
+from merlin.models.tf.core.combinators import ParallelBlock, ResidualBlock, SequentialBlock
+from merlin.models.tf.core.index import IndexBlock, TopKIndexBlock
+from merlin.models.tf.core.inputs import InputBlock
+from merlin.models.tf.core.masking import CausalLanguageModeling, MaskedLanguageModeling
+from merlin.models.tf.core.tabular import AsTabular, Filter, TabularBlock
+from merlin.models.tf.core.transformations import (
+    AsDenseFeatures,
+    AsSparseFeatures,
+    CategoricalOneHot,
+    ExpandDims,
+    LabelToOneHot,
+    StochasticSwapNoise,
+)
 from merlin.models.tf.dataset import sample_batch
 from merlin.models.tf.features.continuous import ContinuousFeatures
 from merlin.models.tf.features.embedding import (
