+++ merlin/models/tf/core/aggregation.py
@@ -21,8 +21,8 @@ import tensorflow as tf
 from tensorflow.python.keras.layers import Dot
 
 from merlin.models.config.schema import requires_schema
-from merlin.models.tf.blocks.core.base import Block
-from merlin.models.tf.blocks.core.tabular import TabularAggregation
+from merlin.models.tf.core.base import Block
+from merlin.models.tf.core.tabular import TabularAggregation
 from merlin.models.tf.typing import TabularData
 from merlin.models.tf.utils import tf_utils
 from merlin.models.utils.schema_utils import schema_to_tensorflow_metadata_json
