+++ merlin/models/tf/prediction_tasks/next_item.py
@@ -19,17 +19,17 @@ from typing import Optional
 import tensorflow as tf
 from tensorflow.python.layers.base import Layer
 
-from merlin.models.tf.blocks.core.base import Block
-from merlin.models.tf.blocks.core.masking import MaskingHead
-from merlin.models.tf.blocks.core.transformations import (
+from merlin.models.tf.blocks.retrieval.base import ItemRetrievalScorer
+from merlin.models.tf.blocks.sampling.cross_batch import PopularityBasedSampler
+from merlin.models.tf.core.base import Block
+from merlin.models.tf.core.masking import MaskingHead
+from merlin.models.tf.core.transformations import (
     ItemsPredictionWeightTying,
     L2Norm,
     LabelToOneHot,
     LogitsTemperatureScaler,
     RemovePad3D,
 )
-from merlin.models.tf.blocks.retrieval.base import ItemRetrievalScorer
-from merlin.models.tf.blocks.sampling.cross_batch import PopularityBasedSampler
 from merlin.models.tf.losses.base import LossType
 from merlin.models.tf.metrics.ranking import ranking_metrics
 from merlin.models.tf.prediction_tasks.classification import (
